+++ harfbuzz/src/hb-ot-shape-complex-arabic.cc
@@ -297,7 +297,7 @@ arabic_joining (hb_buffer_t *buffer)
   /* Check pre-context */
   for (unsigned int i = 0; i < buffer->context_len[0]; i++)
   {
-    unsigned int this_type = get_joining_type (buffer->context[0][i], buffer->unicode->general_category (buffer->context[0][i]));
+    unsigned int this_type = get_joining_type (buffer->context[0][i], hb_ucd_general_category (buffer->context[0][i]));
 
     if (unlikely (this_type == JOINING_TYPE_T))
       continue;
@@ -332,7 +332,7 @@ arabic_joining (hb_buffer_t *buffer)
 
   for (unsigned int i = 0; i < buffer->context_len[1]; i++)
   {
-    unsigned int this_type = get_joining_type (buffer->context[1][i], buffer->unicode->general_category (buffer->context[1][i]));
+    unsigned int this_type = get_joining_type (buffer->context[1][i], hb_ucd_general_category (buffer->context[1][i]));
 
     if (unlikely (this_type == JOINING_TYPE_T))
       continue;