+++ harfbuzz/src/hb-ot-shape-complex-khmer.cc
@@ -422,7 +422,7 @@ decompose_khmer (const hb_ot_shape_normalize_context_t *c,
     case 0x17C5u  : *a = 0x17C1u; *b= 0x17C5u; return true;
   }
 
-  return (bool) c->unicode->decompose (ab, a, b);
+  return (bool) hb_ucd_decompose (ab, a, b);
 }
 
 static bool
@@ -432,10 +432,10 @@ compose_khmer (const hb_ot_shape_normalize_context_t *c,
 	       hb_codepoint_t *ab)
 {
   /* Avoid recomposing split matras. */
-  if (HB_UNICODE_GENERAL_CATEGORY_IS_MARK (c->unicode->general_category (a)))
+  if (HB_UNICODE_GENERAL_CATEGORY_IS_MARK (hb_ucd_general_category (a)))
     return false;
 
-  return (bool) c->unicode->compose (a, b, ab);
+  return (bool) hb_ucd_compose (a, b, ab);
 }