+++ harfbuzz/src/hb-ot-shape-complex-use.cc
@@ -573,10 +573,10 @@ compose_use (const hb_ot_shape_normalize_context_t *c,
 	     hb_codepoint_t *ab)
 {
   /* Avoid recomposing split matras. */
-  if (HB_UNICODE_GENERAL_CATEGORY_IS_MARK (c->unicode->general_category (a)))
+  if (HB_UNICODE_GENERAL_CATEGORY_IS_MARK (hb_ucd_general_category (a)))
     return false;
 
-  return (bool)c->unicode->compose (a, b, ab);
+  return (bool)hb_ucd_compose (a, b, ab);
 }