+++ CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - All external dependencies. No icu, glib, freetype, etc.
 - `hb_shape_list_shapers`, since we only have `ot` and `fallback`.
 - Deprecated methods.
+- `hb_unicode_funcs_t`, since we are using only internal Unicode tables.